1.1 restructured the code
